+++ packages/deployed-backend-client/src/deployed_backend_client.test.ts
@@ -28,6 +28,8 @@ import { AmplifyClient } from '@aws-sdk/client-amplify';
 import { GetObjectCommand, S3 } from '@aws-sdk/client-s3';
 import { DeployedResourcesEnumerator } from './deployed-backend-client/deployed_resources_enumerator.js';
 import { StackStatusMapper } from './deployed-backend-client/stack_status_mapper.js';
+import { ArnGenerator } from './deployed-backend-client/arn_generator.js';
+import { AccountIdParser } from './deployed-backend-client/account_id_parser.js';
 
 const listStacksMock = {
   NextToken: undefined,
@@ -209,8 +211,13 @@ void describe('Deployed Backend Client', () => {
 
     cfnClientSendMock.mock.mockImplementation(mockSend);
 
+    const arnGeneratorMock = new ArnGenerator();
+    const accountIdParserMock = new AccountIdParser();
+    mock.method(arnGeneratorMock, 'generateArn', () => undefined);
     const deployedResourcesEnumerator = new DeployedResourcesEnumerator(
-      new StackStatusMapper()
+      new StackStatusMapper(),
+      arnGeneratorMock,
+      accountIdParserMock
     );
     mock.method(deployedResourcesEnumerator, 'listDeployedResources', () => []);
 
@@ -219,7 +226,8 @@ void describe('Deployed Backend Client', () => {
       mockS3Client,
       mockBackendOutputClient,
       deployedResourcesEnumerator,
-      new StackStatusMapper()
+      new StackStatusMapper(),
+      accountIdParserMock
     );
   });
 
@@ -359,8 +367,13 @@ void describe('Deployed Backend Client pagination', () => {
     };
 
     cfnClientSendMock.mock.mockImplementation(mockSend);
+    const arnGeneratorMock = new ArnGenerator();
+    const accountIdParserMock = new AccountIdParser();
+    mock.method(arnGeneratorMock, 'generateArn', () => undefined);
     const deployedResourcesEnumerator = new DeployedResourcesEnumerator(
-      new StackStatusMapper()
+      new StackStatusMapper(),
+      arnGeneratorMock,
+      accountIdParserMock
     );
     mock.method(deployedResourcesEnumerator, 'listDeployedResources', () => []);
 
@@ -369,7 +382,8 @@ void describe('Deployed Backend Client pagination', () => {
       mockS3Client,
       mockBackendOutputClient,
       deployedResourcesEnumerator,
-      new StackStatusMapper()
+      new StackStatusMapper(),
+      accountIdParserMock
     );
   });